// styles/styled-system/types.ts
export interface TypographyProps
  extends ITypographyProps,
    SpaceProps,
    TextAlignProps,
    ColorProps {
  textTransform?: ResponsiveValue<CSS.Property.TextTransform>
  fontWeight?: ResponsiveValue<FontWeight>
  whiteSpace?: ResponsiveValue<CSS.Property.WhiteSpace>
  wordBreak?: ResponsiveValue<CSS.Property.WordBreak>
  textDecoration?: ResponsiveValue<CSS.Property.TextDecoration>
}

export type SpaceProps = Pick<
  _SpaceProps,
  | "m"
  | "mt"
  | "mr"
  | "mb"
  | "ml"
  | "mx"
  | "my"
  | "p"
  | "pt"
  | "pr"
  | "pb"
  | "pl"
  | "px"
  | "py"
>

export interface TransitionProps {
  transition?: ResponsiveValue<string>
}

export interface OverflowWrapProps {
  overflowWrap?: ResponsiveValue<CSS.Property.OverflowWrap>
}
export interface WordBreakProps {
  wordBreak?: ResponsiveValue<CSS.Property.WordBreak>
}
export interface VisibilityProps {
  visibility?: CSS.Property.Visibility
}

export interface ColorProps {
  textColor?: ResponsiveValue<ColorKeys | string>
  backgroundColor?: ResponsiveValue<ColorKeys | string>
}